# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Movieflix API",
		"description": "API responsável por gerenciar os filmes do movieflix",
		"version": "1.0.0"
	},
	"paths": {
		"/movies": {
			"get": {
				"tags": ["Movie"],
				"summary": "Obter todos os filmes",
				"responses": {
					"200": {
						"description": "Retorna uma lista de filmes",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Movie"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Movie"],
				"summary": "Inserir um novo filme",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MovieInput"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Filme cadastrado com sucesso"
					},
					"409": {
						"description": "Já existe um filme cadastrado com esse título"
					},
					"500": {
						"description": "Falha ao cadastrar um filme"
					}
				}
			}
		},
		"/movies/{id}": {
			"put": {
				"tags": ["Movie"],
				"summary": "Atualizar um filme",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "ID do filme a ser atualizado",
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MovieInput"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Filme atualizado com sucesso"
					},
					"404": {
						"description": "Filme não encontrado"
					},
					"500": {
						"description": "Falha ao atualizar o registro do filme"
					}
				}
			},
			"delete": {
				"tags": ["Movie"],
				"summary": "Remover um filme",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "ID do filme a ser removido",
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Filme removido com sucesso"
					},
					"404": {
						"description": "Filme não encontrado"
					},
					"500": {
						"description": "Falha ao remover o registro"
					}
				}
			}
		},
		"/movies/{genreName}": {
			"get": {
				"tags": ["Movie"],
				"summary": "Filtrar filmes por gênero",
				"parameters": [
					{
						"name": "genreName",
						"in": "path",
						"required": true,
						"description": "Nome do gênero a ser pesquisado",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Movie"
									}
								}
							}
						}

					},
					"500": {
						"description": "Falha ao filtrar filmes por gênero"
					}
				}
			}
		},		
		"/genres/{id}": {
			"put": {
				"tags": ["Genre"],
				"summary": "Atualizar um gênero",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "ID do gênero a ser atualizado",
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenreInput"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Gênero atualizado com sucesso!"
					},
					"404": {
						"description": "Gênero não encontrado"
					},
					"500": {
						"description": "Falha ao tentar a atualizar os dados desde gênero!"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Movie": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"description": "ID do filme"
					},
					"title": {
						"type": "string",
						"description": "Título do filme"
					},
					"genre_id": {
						"type": "integer",
						"description": "Gênero do filme"
					},
					"genres": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer",
								"description": "ID do gênero"
							},
							"name": {
								"type": "string",
								"description": "Nome do gênero"
							}
						}
					},
					"language_id": {
						"type": "integer",
						"description": "Idioma do filme"
					},
					"languages": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer",
								"description": "ID do idioma"
							},
							"name": {
								"type": "string",
								"description": "Nome do idioma"
							}
						}
					},
					"oscar_count": {
						"type": "integer",
						"description": "Quantidade de Oscars que o filme já ganhou"
					},
					"release_date": {
						"type": "string",
						"format": "date",
						"description": "Data de lançamento do filme"
					}
				}
			},
			"MovieInput": {
				"type":"object",
				"required": [
					"title",
					"genre_id",
					"language_id",
					"oscar_count",
					"release_date"
				],
				"properties": {
					"title": {
						"type": "string",
						"description": "Título do filme"
					},
					"genre_id": {
						"type":"integer",
						"description": "ID do gênero do filme"
					},
					"language_id": {
						"type":"integer",
						"description":"ID do idioma do filme"
					},
					"oscar_count": {
						"type":"integer",
						"description":"Quantidade de Oscars que o filme já ganhou"
					},
					"release_date": {
						"type": "string",
						"format": "date",
						"description": "Data de lançamento do filme"
					}
				}
			},
			"GenreInput": {
				"type":"object",
				"required": [
					"id",
					"name"
				],
				"properties": {					
					"name": {
						"type": "string",
						"description": "Nome do gênero"
					}
				}
			}
		}
		
	}
}